fix(shell-api): implement own map/forEach for cursors MONGOSH-936
Because the driver expects `.map()` and `.forEach()` callbacks to work
synchronously, but mongosh users can use pseudo-synchronous async
functions, the driver cannot properly handle the functions we pass
to it.
In order to fix this, if we don’t jump through a lot of extra hoops,
it’s easiest to just implement `.map()` and `.forEach()` ourselves.